Actualizado editar usuario

+++ src/java/com/example/dao/ExcelDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface ExcelDAO  extends AutoCloseable {
     public void create(Estudiante est) throws DAOException;//CREATE
-    public Estudiante findByCedula(int cedula) throws DAOException;//READ
+    public Estudiante findByCedula(long cedula) throws DAOException;//READ
     public void update(Estudiante est) throws DAOException;//UPDATE
     public void delete(int cedula) throws DAOException;//DELETE
     public List<Estudiante> getAllEstudiantes() throws DAOException;
